\name{aids}
\alias{aids}
\title{AIDS Incubation}
\description{
    A data set collected by Centers for Disease Control and Prevention
    concerning AIDS patients who were infected with the HIV virus
    through blood transfusion.
}
\usage{data(aids)}
\format{
    A data frame containing 295 observations on the following variables.
    \tabular{ll}{
	\code{incu} \tab Time from HIV infection to AIDS diagnosis.\cr
	\code{infe} \tab Time from HIV infection to end of data
	collection (July 1986).\cr
	\code{age} \tab Age at time of blood transfusion.
    }
}
\source{
    Wang, M.-C. (1989), A semiparametric model for randomly truncated
    data.  \emph{Journal of the American Statistical Association},
    \bold{84}, 742--748. 
}
\keyword{datasets}
