\name{buffalo}
\alias{buffalo}
\title{Buffalo Annual Snowfall}
\description{
    Annual snowfall accumulations in Buffalo, NY from 1910 to 1973.
}
\usage{data(buffalo)}
\format{
    A vector of 63 numerical values.
}
\source{
    Scott, D. W. (1985), Average shifted histograms: Effective
    nonparametric density estimators in several dimensions.  \emph{The
    Annals of Statistics}, \bold{13}, 1024--1040.
}
\keyword{datasets}
