\name{clim}
\alias{clim}
\title{Average Temperatures During December 1980 Through February 1981}
\description{
    Average temperatures at 690 weather stations during December 1980
    through February 1981.
}
\usage{data(clim)}
\format{
    A data frame containing 690 observations on the following variables.
    \tabular{ll}{
	\code{temp} \tab Average temperature, in Celsius.\cr
	\code{geog} \tab Geographic location (latitude,longitude), in
	degrees, as a matrix.
    }
}
\source{
    This is reformulated from the data frame \code{climate} in the R
    package \code{assist} by Yuedong Wang and Chunlei Ke.
}
\keyword{datasets}
