\name{eyetrack}
\alias{eyetrack}
\title{Eyesight Fixation in Eyetracking Experiments}
\description{
    Eyesight Fixation during some eyetracking experiments in liguistic
    studies.
}
\usage{data(eyetrack)}
\format{
    A data frame containing 13891 observations on the following variables.
    \tabular{ll}{
	\code{time} \tab Time, in ms.\cr
	\code{color} \tab Binary indicator, 1 if eyesight fixed on target
	or color competitor, a factor.\cr
	\code{object} \tab Binary indicator, 1 if eyesight fixed on target
	or object competitor, a factor.\cr
	\code{id} \tab Identification of homogeneous sessions, a factor.\cr
	\code{cnt} \tab Multiplicity count.
    }
}
\source{
    Dr. Anouschka Foltz.
}
\references{
    Gu, C. and Ma, P. (2011), Nonparametric regression with
    cross-classified responses.  Manuscript.
}
\keyword{datasets}
