\name{nox}
\alias{nox}
\title{NOx in Engine Exhaust}
\description{
    Data from an experiment in which a single-cylinder engine was run
    with ethanol to see how the NOx concentration in the exhaust
    depended on the compression ratio and the equivalence ratio.
}
\usage{data(nox)}
\format{
    A data frame containing 88 observations on the following variables.
    \tabular{ll}{
	\code{nox} \tab NOx concentration in exhaust.\cr
	\code{comp} \tab Compression ratio.\cr
	\code{equi} \tab Equivalence ratio.
    }
}
\source{
    Brinkman, N. D. (1981).  Ethanol fuel -- a single-cylinder engine
    study of efficiency and exhaust emissions.  \emph{SAE Transactions},
    \bold{90}, 1410--1424.
}
\references{
    Cleveland, W. S. and Devlin, S. J. (1988).  Locally weighted
    regression: An approach to regression analysis by local fitting.
    \emph{Journal of the American Statistical Association}, \bold{83},
    596--610.

    Breiman, L. (1991).  The pi method for estimating multivariate
    functions from noisy data.  \emph{Technometrics}, \bold{33},
    125--160. 
}
\keyword{datasets}
