\name{penny}
\alias{penny}
\title{Thickness of US Lincoln Pennies}
\description{
    Thickness of US Lincoln pennies minted during years 1945 through
    1989.
}
\usage{data(nox)}
\format{
    A data frame containing 90 observations on the following variables.
    \tabular{ll}{
	\code{year} \tab Year minted.\cr
	\code{mil} \tab Thickness in mils.
    }
}
\source{
    Scott, D. W. (1992), \emph{Multivariate Density Estimation: Theory,
    Practice and Visualization}.  New York: Wiley.
}
\references{
    Gu, C. (1995), Smoothing spline density estimation: Conditional
    distribution, \emph{Statistica Sinica}, \bold{5}, 709--726.
    
    Scott, D. W. (1992), \emph{Multivariate Density Estimation: Theory,
    Practice and Visualization}.  New York: Wiley.
}
\keyword{datasets}
