\name{Sachs}
\alias{Sachs}
\title{Protein Expression in Human Immune System Cells}
\description{
    Data concerning protein expression levels in human immune system
    cells under stimulations.
}
\usage{data(Sachs)}
\format{
    A data frame containing 7466 cells with expression levels of the
    following proteins, on the \code{log10} scale of the original.
    \tabular{ll}{
        \code{praf}.\cr
        \code{pmek}.\cr
	\code{plcg}.\cr
	\code{pip2}.\cr
	\code{pipp3}.\cr
	\code{p44.42}.\cr
	\code{pakts473}.\cr
	\code{pka}.\cr
	\code{pkc}.\cr
	\code{p38}.\cr
	\code{pjnk}.\cr
    }
}
\references{
    Sachs, K., Perez, O., Pe'er, D., Lauffenburger, D. A., and Nolan,
    G. P. (2005), Causal protein-signaling networks derived from
    multiparameter single-cell data. \emph{Science}, \bold{308 (5732)},
    523--529.
}
\keyword{datasets}
