\name{summary.sscopu}
\alias{summary.sscopu}
\title{Calculating Kendall's Tau and Spearman's Rho for 2-D Copula
    Density Estimates}
\description{
    Calculate Kendall's tau and Spearman's rho for 2-D copula density
    estimates.
}
\usage{
\method{summary}{sscopu}(object, ...)
}
\arguments{
    \item{object}{Object of class \code{"sscopu"}.}
    \item{...}{Ignored.}
}
\value{
    A list containing Kendall's tau and Spearman's rho.
}
\seealso{
    Fitting functions \code{\link{sscopu}} and \code{\link{sscopu2}}.
}
\keyword{models}
\keyword{distribution}
\keyword{smooth}
