\name{wesdr}
\alias{wesdr}
\title{Progression of Diabetic Retinopathy}
\description{
    Data derived from the Wisconsin Epidemiological Study of Diabetic
    Retinopathy.
}
\usage{data(wesdr)}
\format{
    A data frame containing 669 observations on the following variables.
    \tabular{ll}{
	\code{dur} \tab Duration of diabetes at baseline, in years.\cr
	\code{gly} \tab Percent of glycosylated hemoglobin at baseline.\cr
	\code{bmi} \tab Body mass index at baseline.\cr
	\code{ret} \tab Binary indicator of retinopathy progression at
	first follow-up.
    }
}
\source{
    Wang, Y. (1997), GRKPACK: Fitting smoothing spline ANOVA models for
    exponential families.  \emph{Communications in Statistics --
    Simulations and Computation}, \bold{26}, 765--782.
}
\references{
    Klein, R., Klein, B. E. K., Moss, S. E., Davis, M. D., and DeMets,
    D. L. (1988), Glycosylated hemoglobin predicts the incidence and
    progression of diabetic retinopathy.  \emph{Journal of the American
    Medical Association}, \bold{260}, 2864--2871.

    Klein, R., Klein, B. E. K., Moss, S. E., Davis, M. D., and DeMets,
    D. L. (1989), The Wisconsin Epidemiologic Study of Diabetic
    Retinopathy. X. Four incidence and progression of diabetic
    retinopathy when age at diagnosis is 30 or more years.
    \emph{Archive Ophthalmology}, \bold{107}, 244--249.

    Wahba, G., Wang, Y., Gu, C., Klein, R., and Klein, B. E. K. (1995),
    Smoothing spline ANOVA for exponential families, with application to
    the Wisconsin Epidemiological Study of Diabetic
    Retinopathy. \emph{The Annals of Statistics}, \bold{23}, 1865--1895.
}
\keyword{datasets}
