\name{wesdr1}
\alias{wesdr1}
\title{Stages of Diabetic Retinopathy}
\description{
    Data derived from the Wisconsin Epidemiological Study of Diabetic
    Retinopathy.
}
\usage{data(wesdr1)}
\format{
    A data frame containing 2049 observations on the following variables.
    \tabular{ll}{
	\code{age} \tab Age of patient.\cr
	\code{dur} \tab Duration of diabetes, in years.\cr
	\code{gly} \tab Percent of glycosylated hemoglobin.\cr
	\code{upro} \tab Ordinal urine protein level.\cr
	\code{insl} \tab Binary indicator of insulin usage.\cr
	\code{ret1} \tab Ordinal retinopathy stage, right eye.\cr
	\code{ret2} \tab Ordinal retinopathy stage, left eye.
    }
}
\keyword{datasets}
