\name{print}
\alias{print.ssanova}
\alias{print.ssanova0}
\alias{print.gssanova}
\alias{print.ssden}
\alias{print.sshzd}
\alias{print.summary.ssanova}
\alias{print.summary.gssanova}
\alias{print.summary.gssanova0}
\title{Print Functions for Smoothing Spline ANOVA Models}
\description{
    Print functions for Smoothing Spline ANOVA models.
}
\usage{
print.ssanova(x, ...)
print.ssanova0(x, ...)
print.gssanova(x, ...)
print.ssden(x, ...)
print.sshzd(x, ...)
print.summary.ssanova(x, digits=6, ...)
print.summary.gssanova(x, digits=6, ...)
print.summary.gssanova0(x, digits=6, ...)
}
\arguments{
    \item{x}{Object of class \code{ssanova}, \code{summary.ssanova},
        \code{summary.gssanova}, or \code{ssden}.}
    \item{digits}{Number of significant digits to be printed in
	values.}
    \item{...}{Ignored.}
}
\seealso{
    \code{\link{ssanova}}, \code{\link{ssanova0}},
    \code{\link{gssanova}}, \code{\link{gssanova0}},
    \code{\link{ssden}}, \code{\link{sshzd}},
    \code{\link{summary.ssanova}}, \code{\link{summary.gssanova}},
    \code{\link{summary.gssanova0}}.
}
\keyword{smooth}
\keyword{models}
