\name{print.ssanova}
\alias{print.summary.ssanova}
\alias{print.summary.gssanova}
\title{Print Functions for Smoothing Spline ANOVA Models}
\description{
    Print functions for Smoothing Spline ANOVA models.
}
\usage{
print.ssanova(obj)
print.summary.ssanova(obj, digits=6)
print.summary.gssanova(obj, digits=6)
}
\arguments{
    \item{obj}{Object of class \code{ssanova}, \code{summary.ssanova},
        or \code{summary.gssanova}.}
    \item{digits}{Number of significant digits to be printed in values.}
}
\seealso{
    \code{\link{ssanova}}, \code{\link{gssanova}},
    \code{\link{summary.ssanova}}, \code{\link{summary.gssanova}}.
}
\keyword{smooth}
