\name{print.ssanova}
\alias{print.summary.ssanova}
\alias{print.summary.gssanova}
\title{Print Functions for Smoothing Spline ANOVA Models}
\description{
    Print functions for Smoothing Spline ANOVA models.
}
\usage{
print[.ssanova](obj)
print[.summary.ssanova](obj)
print[.summary.gssanova](obj)
}
\arguments{
    \item{obj}{Object of class \code{ssanova}, \code{summary.ssanova},
	or \code{summary.gssanova}.}
}
\seealso{
    \code{\link{ssanova}}, \code{\link{gssanova}},
    \code{\link{summary.ssanova}}, \code{\link{summary.gssanova}}.
}
\keyword{smooth}
