\name{Sachs}
\alias{Sachs}
\title{Protein Expression in Human Immune System Cells}
\description{
    Data concerning protein expression levels in human immune system
    cells under stimulations.
}
\usage{data(Sachs)}
\format{
    A data frame containing 7466 cells, with flow cytometry measurements
    of 11 phosphorylated proteins and phospholipids, on the \code{log10}
    scale of the original.
    \tabular{ll}{
        \code{praf} \tab Raf phosphorylated at S259.\cr
        \code{pmek} \tab Mek1/mek2 phosphorylated at S217/S221.\cr
	\code{plcg} \tab Phosphorylation of phospholipase \eqn{C-\gamma}
	on Y783.\cr
	\code{pip2} \tab Phophatidylinositol 4,5-biphosphate.\cr
	\code{pip3} \tab Phophatidylinositol 3,4,5-triphosphate.\cr
	\code{p44.42} \tab Erk1/erk2 phosphorylated at T202/Y204.\cr
	\code{pakts473} \tab AKT phosphorylated at S473.\cr
	\code{pka} \tab Phosphorylation of of protein kinase A
	substrates on 3 sites.\cr
	\code{pkc} \tab Phosphorylation of of protein kinase C
	substrates on S660.\cr
	\code{p38} \tab Erk1/erk2 phosphorylated at T180/Y182.\cr
	\code{pjnk} \tab Erk1/erk2 phosphorylated at T183/Y185.\cr
    }
}
\source{
    Sachs, K., Perez, O., Pe'er, D., Lauffenburger, D. A., and Nolan,
    G. P. (2005), Causal protein-signaling networks derived from
    multiparameter single-cell data. \emph{Science}, \bold{308 (5732)},
    523--529.
}
\keyword{datasets}
